+++ clox/src/debug.h
@@ -3,6 +3,6 @@
 #include "bytecode_chunk.h"
 
 void disassemble_chunk(bytecode_chunk* chunk, const char* name);
-int disassemble_instruction(bytecode_chunk* chunk, int offset);
+size_t disassemble_instruction(bytecode_chunk* chunk, size_t offset);
 
 #endif
